requirements.txt
annotated-types==0.7.0
bcrypt==4.3.0
cffi==1.17.1
coverage==7.8.0
cryptography==45.0.2
iniconfig==2.1.0
packaging==25.0
paramiko==3.5.1
pluggy==1.6.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyNaCl==1.5.0
pytest==8.3.5
pytest-cov==6.1.1
python-dotenv==1.1.0
PyYAML==6.0.2
tenacity==9.1.2
types-paramiko==3.5.0.20250516
types-PyYAML==6.0.12.20250516
typing-inspection==0.4.0
typing_extensions==4.13.2
